+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, following t
 -   Renamed item/items to datum/data in several places
 -   Can manipulate markers via `hm.extensions.marker?.drawMarkers({...})`
 -   Default coloring is "everything gray", even for numeric values
+-   Color scales are created by `ColorScale.continuous`
 -   Dropped `formatDataItem` function
 
 ## [0.9.0] - 2024-04-29